pyproject.toml
[tool.poetry]
name = "imager"
version = "0.0.1"
description = "IMager is a telegram bot project that creates images from other images"
authors = ["LittlePemp <52671349+LittlePemp@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.21"
Pillow = "^9.2.0"
transliterate = "^1.10.2"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
bs4 = "^0.0.1"
isort = "^5.10.1"
pydirectory = "^0.1.15"
PyDictionary = "^2.0.1"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
